requirements.txt
spotipy==2.24.0
python-dotenv==1.0.1
requests==2.32.3
click==8.1.7
rich==13.9.4
Pillow==12.2.0
